# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: RESTful APIs of Material Management App
host: localhost:3000
basePath: /api
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and retourned as answer of a call
      to the route /login giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

          Bearer xxxxxx.yyyyyyy.zzzzzz
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /uploadSupplementaryDoc:
    x-swagger-router-controller: upload-controller
    post:
      tags:
        - Supplementary Document File
      operationId: uploadSupplementaryDoc
      consumes:
        - multipart/form-data
      summary: "Upload a Supplementary Document"
      description: "Upload the supplementary document before crraeting backfill/disposal requests"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - in: "formData"
          name: "supplementary_file"
          description: "Supplementary Document"
          required: true
          type: file
          x-mimetype: image/png, image/jpeg, application/pdf
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/Resource"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
  /deleteSupplementaryDoc:
    x-swagger-router-controller: upload-controller
    post:
      tags:
        - Supplementary Document File
      operationId: deleteSupplementaryDoc
      summary: "Delete Supplementary Document"
      description: "Remove a supplementary document by a given id"
      security:
        - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: fileDeletion
          in: body
          required: true
          schema:
            $ref: "#/definitions/DocumentDeletion"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/Success"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Resource:
    type: object
    properties:
      resource:
        type: string
...
  DocumentDeletion:
    type: object
    properties:
      supplementary_document:
        type: string
...
  Success:
    type: object
    properties:
      status:
        type: boolean
      message:
        type: string
  Error:
    type: object
    properties:
      message:
        type: string